move DefaultTorSettings class to settings module
