+++ topl-android/src/main/java/io/matthewnelson/topl_android/OnionProxyContext.kt
@@ -13,7 +13,7 @@ See the Apache 2 License for the specific language governing permissions and lim
 package io.matthewnelson.topl_android
 
 import android.os.Process
-import io.matthewnelson.topl_android.settings.DefaultTorSettings
+import io.matthewnelson.topl_android_settings.DefaultTorSettings
 import io.matthewnelson.topl_android_settings.TorSettings
 import io.matthewnelson.topl_android.settings.TorSettingsBuilder
 import io.matthewnelson.topl_android.util.FileUtilities
